P-ROC: update for new yaml-cpp API P-ROC: quiet a compiler warning P-ROC: cleanup; make code conditional on coreGlobals.p_rocEn P-ROC: move "\n" from start to end of printf() P-ROC: allow custom flipper switch rules  Allows support of Demolition Man's unique flipper configuration  (multiple buttons trigger multiple flippers). wpc/dm: add lamp positions to Demo Man wpc/dm: extend simulator to all shots but cryoclaw wpc/dm: simulate elevator/claw, enable Buy-In
patches by tomlogic!

git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@5058 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/p-roc/proc_shift_reg.c
@@ -73,7 +73,7 @@ void proc_shiftRegEnqueue(int bit) {
     shr.bit -= 32;
   }
   if (bit)
-    shr.queue |= (1UL << shr.bit);
+    shr.queue |= ((UINT64)1 << shr.bit);
   ++shr.bit;
   
   if (!shr.timerEnabled) {
